Rename package name to kodi_addon_checker
