+++ tests/test_common.py
@@ -5,7 +5,7 @@ import os
 import json
 import pytest
 from PIL import Image
-from addon_checker.common import check_config, has_transparency
+from kodi_addon_checker.common import check_config, has_transparency
 
 FIXTURE_PATH = os.path.join("tests", "fixtures")
